requirements.txt
Flask
Flask_Cors
gradio>=3.7.0
numpy==1.23.0
pyworld==0.2.5
scipy==1.10.0
SoundFile==0.12.1
paddlepaddle-gpu==1.13.1
paddlepaddle-gpuaudio==0.13.1
paddlepaddle-gpucrepe
tqdm
scikit-maad
praat-parselmouth
onnx
onnxsim
onnxoptimizer
fairseq==0.12.2
librosa==0.9.1
tensorboard
tensorboardX
edge_tts
